1.0.7

+++ src/main/java/com/cpu/web/repository/BulletinRepository.java
@@ -3,5 +3,5 @@ package com.cpu.web.repository;
 import com.cpu.web.entity.Bulletin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BulletinRepository extends JpaRepository<Bulletin, Integer> {
+public interface BulletinRepository extends JpaRepository<Bulletin, Long> {
 }
